+++ aws/EC2/getAllEC2Regions.ts
