+++ aws/EC2/getInstanceByName.ts
