+++ cli/utils/promptUserForRegion.ts
