+++ cli/utils/setupAws.ts
