+++ cli/utils/spinner.ts
